requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# NLP and ML
transformers==4.37.2
sentence-transformers==2.3.1
torch==2.1.2
accelerate==0.26.1

# Vector store and retrieval
faiss-cpu==1.7.4
langchain==0.1.4
langchain-community==0.0.16

# Document processing
pypdf==4.0.1
python-docx==1.1.0

# Translation
googletrans==4.0.0rc1
indicnlp==0.91

# Utilities
numpy==1.26.3
pandas==2.2.0
pyyaml==6.0.1
tqdm==4.66.1
python-dotenv==1.0.0